Add metadata.txt file with raw compression and pre-dexing
Summary: Same as previous diff except this time for when we are doing pre-dexing. When we are doing pre-dexing each line of the `metadata.txt` is collected from the `d8` action and then we assemble them at the end. We were already doing that for `jar` compression, just do the same for `raw` compression.

Reviewed By: ianlevesque

Differential Revision: D38742807

fbshipit-source-id: b65047c800d15efb78fcfdf07254839cd8930ca6