Return a XcodeDataInfo provider such that it can be read in BXL
Summary:
In the next diff, we use BXL to work around the fact that we can't get implied/direct deps of a target in the rules.

For the BXL script, we want to get the same information we write in the `xcode-data` subtarget. However, buck2 to write files is slower, compared to getting the information directly from providers after analysis.

This requires us to return a provider for the rules that we want [xcode-data] support.

For backwards compatibility (with tests and project generator), in addition to defining a subtarget, return a XcodeDataInfo provider. In the future, we will likely be able to get rid of the subtarget.

Note, for apple_test, we want to get just the XcodeDataInfo provider. We cannot do type checks on providers, so I need to return it seperately in `_CxxLibraryParameterizedOutput`

Reviewed By: milend

Differential Revision: D38833535

fbshipit-source-id: 57c38e7596592154383e6b59d5a176890da9dfc0