Merge TargetMerger into TargetNode
Summary:
TargetMerger and TargetNode do much of the same thing, and there is a lot of duplicated code/logic. Specifically, TargetMerger takes targets to merge, computes the aggregated files and then creates a TargetNode.

In the next few diffs, we're going to start making changes that will need updating in both unless the code is deduped and merged. Do that now.

There is only one functional difference. TargetMerger created `targets_to_prebuild` which included the `argsfiles` subtargets for the targets with files to compile. This data was combined with generated files, etc. These argsfiles are now built with the other argsfiles, so the runtime numbers are a bit different.

Before:
```
           sources: 15,398
           headers: 4,908
  exported headers: 23,314
       other files: 5,449
       total files: 49,069
  prebuild targets: 19,919
```

After:
```
           sources: 15,398
           headers: 4,908
  exported headers: 23,314
       other files: 5,449
       total files: 49,069
  prebuild targets: 14,375
```

The difference is in the targets with argsfiles.

Reviewed By: milend

Differential Revision: D38851934

fbshipit-source-id: b77f66bd9e75d3658b1de06aa6b17bcc8fbf548d