buck2_server_ctx crate
Summary:
Parts of `buck2_server` crate: `ServerCommandContext` and a few functions.

This is needed to move `test` and `bxl` command implementations to corresponding crates, `buck2_test` (D38849510) and `buck2_bxl` (D38851147).

`buck2_server_ctx` does not depend on `buck2_build_api`, so it does not slow down compilation.

Dependency graph with the top of the stack (D38851147):
{F762222810}

Reviewed By: scottcao

Differential Revision: D38849511

fbshipit-source-id: 7dcf716ca3ee8102111c9a5191592daf8b6ac1fd