Disable fat platforms on linux-aarch64
Summary:
Since we're enabling using x86 RE workers, we're getting into situations
where we're trying to build a PAR locally that targets the fat platform
and failing since the Python toolchain we use for it is not cross-build
capable.

I'm sure a better way to do this, so just disabling fat platforms on
linux-aarch64.

https://fb.prod.workplace.com/groups/1858357931081326/posts/3691279957789105/

Reviewed By: ottoni

Differential Revision: D38884784

fbshipit-source-id: b91e84925b6117e4784a43dbad9db328797a70f4