Clean up clang flags for post-link compilation
Summary:
Intercept and cleanup the flags before handing them over to fbcc.

Clang driver can add certain flags that conflict with the existing bitcode attribute. For example, clang driver always tries to add `-triple`, `-target-cpu` and `-tune-cpu` even if they are not explicitly given. This could cause codegen discrepancy if the input bitcode was compiled with different values of these flags. Previously applied values are already encoded in the bitcode file as metadata, and we should just use them.

Differential Revision: D38881137

fbshipit-source-id: 2e5be1e888372b0ebe7cae1829144c0f7d02ce03