buck2: python_binary: allow packaging of cxx_binary with dyanmic dependencies (e.g. from link groups) inside standalone python_binary
Summary:
A `python_binary` can package a `cxx_binary` as a resource.

https://www.internalfb.com/code/fbsource/[89db3c374c74]/fbcode/unicorn/native_indexer/tools/TARGETS?lines=236-254

The `cxx_binary` may use link-groups in which case it has dynamic dependencies inside a symlink tree (eg. for `native_indexer` it's `__native_indexer__shared_libs_symlink_tree`):
https://www.internalfb.com/code/fbsource/[3622cff5139b058fc7ed569207d48891b7e5aedf]/fbcode/unicorn/native_indexer/TARGETS?lines=12-19
https://www.internalfb.com/code/fbsource/[3622cff5139b058fc7ed569207d48891b7e5aedf]/fbcode/unicorn/build/hfsort.bzl?lines=11%2C31-35

The symlink-tree was considered a "hidden" dependency and lead to a build failure:
- this explicit failure was great because without it we would have created invalid python_binary that packaged a cxx_binary without its shared lib dependencies
- explicit failure added in {D34961685}

https://www.internalfb.com/code/fbsource/[89db3c374c74]/fbcode/buck2/prelude/python/python_binary.bzl?lines=234-247

Add the symlink-tree to the manifests (so its contents are packaged inside the python_binary next to the corresponding cxx_binary) and re-classify the symlink-tree dep as a non-hidden dependency.

Reviewed By: christycylee

Differential Revision: D38947009

fbshipit-source-id: ac0798ffbf2252fbd0a794cdc93add94095fb88b