Use correct canary class name for modules
Summary:
Before, we were just using the module name, but in actual fact in buck1 we use a format of `storeXXXX`, where `XXXX` is the hash of the module name.

Not sure if we actually need to replicate that, but it's easiest to make things as similar as possible, so doing that here.

Reviewed By: bchang7

Differential Revision: D38827940

fbshipit-source-id: 5a35163a420767219bc210a3d9d9185e3650e169