Support Voltron when not doing pre-dexing
Summary:
Right now, when we're not doing pre-dexing, we send all of our library jars off to `MultiDexExecutableMain.java`, which runs `d8` on the jars (`d8` is responsible for assigning each jar to a dex) to produce a primary dex and N secondary dexes. We then do any compression necessary on the secondary dexes.

Now, we introduce a dynamic action so that we can read in the APK Module Graph and assign each jar to a module. We then run `MultiDexExecutableMain.java` for each module, and collate the results for all the non-root modules into a symlink tree which is then packaged into the APK.

Reviewed By: bchang7

Differential Revision: D38787230

fbshipit-source-id: 6c09752ecdef4ce1d5dcaf9210f380a6250e28a3