Support xz and xzs for predexed APKs using Voltron config
Summary:
The way that we implement `xz` and `xzs` compression when pre-dexing is enabled is to first create all the raw dex files, and then run a separate action that collects up all those files and compresses them.

We used to do that for just the secondary dexes of the root module. Now, we do it once per module, and stick the resulting output into our symlinked dir (we have to do it this way as we don't know how many modules there are).

Reviewed By: mykola-semko

Differential Revision: D38783401

fbshipit-source-id: 86b0e0967530f4a272a9c41b521fe23e2ecbdd25