Add support for json projections
Summary:
Adds support for json projections. A lot of the core projection handling is shared with arg projections. I considered splitting them, but then all the artifact group stuff also needs to be updated to be aware of each of them and that seemed to not be worth it. There's actually not much interesting here, the support for the values in write_json is pretty trivial and so a lot of this is really just about the handling of the two types of projections.

We could consider dropping entirely the concept of "kind" of projection. Pretty much everything we support today would work fine with a single projection type (that supported being used as both json and cmd_args). The biggest challenge there would be that having a kind allows us to do early validation that the projected values have the right sort of structure to be used in the way that they are claimed to (i.e. we verify the projected values are cmdargs-like or json-valid). Without that early validation, we wouldn't get errors until execution (and we'd likely need to add a bit more complexity to support detecting and propagating the errors at that point).

Reviewed By: krallin

Differential Revision: D38953066

fbshipit-source-id: 9d44e320061ade1b56b508671d4d97a147837747