Always delete cache directory until disk state support is added
Summary:
To support actually using on-disk cache state from buck2, we need to make sure that we never end up reading from a stale cache from disk. krallin pointed out that one way this can happen is if someone checked out an old revision and used a `buck2` that didn't support/understand disk cache (which wouldn't know to delete the cache state). This diff mitigates that possibility by teaching buck2 always delete the buck-out cache directory until on-disk materializer state is supported.
This is one of the two mitigations we are doing to reduce the possibility of this happening.
More context is in the comment block in diff.

Reviewed By: krallin

Differential Revision: D38919531

fbshipit-source-id: 827717ab3432da9326cac84957005f7ccad4fcb9