Allow JUnit tests to depend upon other JUnit test code
Summary:
Somewhat mixed feelings about this, because in an ideal world, we wouldn't allow this. However, this is how things work in `buck1`, and it seems easier just to allow it in `buck2` rather than have to deal with all the "my test passes in buck1 but not in buck2" posts that we'll get if we don't allow it.

Previously, we were exposing the `JavaLibraryInfo` from the test rule, which meant that any test depending upon it was able to build (since that's what we use to construct the compile-time classpath), but it would fail at runtime because we were not exposing `JavaPackagingInfo` (which is what we use to construct the test runtime classpath). Now we expose both, which is consistent.

Reviewed By: krallin

Differential Revision: D39014479

fbshipit-source-id: e3f539798e44ecc12f29bc5e6c1d086f90fabeb1