buck2: caching executor: get the RE use case from the internal executor
Summary:
Like it says in the title. Don't make one up when we instantiate this.

I'm a little torn on the right approach here. One the one hand the design here
is a bit awkward because it means you must have a RE use case for the Local
Executor if you want to layer caching on top.

On the other hand, the alternative is that the CachingExecutor has its own use
case, but that's not ideal either because it creates the possibility of having
a CachingExecutor that wraps a RE Executor that uses a different use case,
which is sort of clowny. Also, it'd be inconsistent with the current
re_platform.

Maybe there's a better design to be had here (I guess this should be hoisted up
to not be properties of any particular executor), but for now this isn't really
making anything much worse and it's pretty simple so let's go with this.

Reviewed By: IanChilds

Differential Revision: D38940384

fbshipit-source-id: 70c3c2a49afba6ae2c879bba31d6a0fb412ce976