dist_thinlto: make final link depends on the index file
Summary:
In current dist_lto implementation, we build a dependencies list ourselves and wrap them with "start-lib/end-lib" in final_link stage,  Which we know may not have the correct order.

with the newly added flag `"-Wl,--thinlto-full-index"` we can ask linker to produce a full list of dependencies in index stage, that can be our source of truth. In this diff we swapped the input for final_link to use the above full index file.

to make it work to fbcode we also did 2 extra things:
1. dist_lto copying files around to work around a buck2 issue (BG: https://fb.workplace.com/groups/buck2users/permalink/3216781525244873/). so we need to modify the index file a little (only for those need opt stage) to reflect the correct location.
2. we use linker wrapper which will add extra dependencies in both index and final_link stage, to avoid double adding we removed these deps from the index files, they will be added back in final_link stage. it's something maybe we can improve in future: T130322878

Reviewed By: apolloww

Differential Revision: D38959206

fbshipit-source-id: 46baa1428f1525cf480779e0d55625114116d79e