Use a dict mapping a label to nodes for link/resource groups
Summary:
Up the stack, we use tsets to replace both the resource graph and linkable graph.

When using tsets, we'll need to use traverse to get all nodes (a reduction which generates a map uses more memory), which returns a list of nodes.

Both link and resource groups need mappings of label -> nodes to process. Previously this was obtained through `graph.nodes`.

To prepare for getting this data from tsets, update the APIs that took a graph object to instead taking a dict of label -> node.

Reviewed By: d16r

Differential Revision: D38954219

fbshipit-source-id: 3c7373a1a138bd0f6fab047b01f9d029dfba01b9