Change matching artifact optimization to only match entry and not deps
Summary:
Currently for matching artifact optimization, we check that `ArtifactValue` is equal, which checks equality on both `entry` and `deps` field of `ArtifactValue`. In reality, we just need to check if `entry` is equal, and we can update `deps` whenever the materialized entry gets redeclared.

This diff does the following:
- Make `ArtifactMaterializationData` hold `entry` and `deps` directly instead of holding just `ArtifactValue`.
- On declare, for matching artifact optimization, just check that `entry` is equal. If `entry` is equal, declare now updates `deps` field of `ArtifactMaterializationData` (instead of doing nothing).

Reviewed By: krallin

Differential Revision: D38136394

fbshipit-source-id: 61b98b56b4bd7e8d64d4feacf856836fd3b23535