Defer computing the graph map until needed
Summary:
Currently, we generate the link and resource graph maps and pass it in as an argument when computing link/resource group maps.

In the case of cxx_libraries which are not linked as a dynamic library (link group), computing the map is wasteful and accounts for ~1.2 GB of allocated memory (total 11.5 GB):

{F763647373}
{F763647374}

By deferring the graph map calculation until needed, we can save most of this allocation.

For both linkable and resource graph, create a wrapper method get_x_map_func which returns a function that gets the map. After total allocated is 9.8 GB.

{F763647810}
{F763647811}

Reviewed By: d16r

Differential Revision: D38980909

fbshipit-source-id: 25303cccc1681d63b56a74a354b4542299f0c7d7