buck2: allow uploading genrules that have cacheable = True
Summary:
Like it says in the title. There is one subtle aspect here, which is that we
partition the cache based on whether this is enabled or not. The goal is to
ensure that if things go wrong only jobs that have local action caching
share their cache.

To do this, if an action may be cached, we set an extra env var in it. This is
all temporary while we roll this out.

#forcetdhashing

Reviewed By: lmvasquezg

Differential Revision: D39132870

fbshipit-source-id: e0feffed9d7247267edf00225d53660056905395