buck2: omnibus: add support for root and exclusion hints
Summary:
Like it says in the title. This adds support for providing hints regarding
what nodes of the graph should be expected to be roots or nodes that should
be expected to be excluded. Exclusions can be provided in two ways:

- Via labels, or simply by default for e.g. cases where we are dealing with a
  cxx python extension, which should obviously be a root (later in this stack
  I made this a bona-fide attribute).
- Via an "environment", which is a list of targets. The goal is to keep this as
  small as possible but it is useful to have in cases where annotating targets
  directly isn't practical.

The end goal is to allow us to expose Omnibus roots that can be shared across
targets, which is being added later in this stack.

When a root is explicitly provided we add it to the LinkableGraph (which is
what makes it a root).

Note: this diff splits up the creation of the Omnibus graph into two tset node,
which is a marginal size increase, but probably doesn't matter (we do the same
thing in c++ linking headers).

#forcetdhashing

Reviewed By: ndmitchell

Differential Revision: D38628579

fbshipit-source-id: 2d35416fa452f79543bab6eb5eef015c9f8cc9e7