buck2: omnibus: add support for sharing roots
Summary:
Like it says in the title. This diff updates the Omnibus logic to do 2 things:

- Produce a sharing candidate when we link something that is known to be a
  root.
- Attempt to reuse the candidate when we need to create a root.

For correctness, the candidate encodes all the assumptions it made when it was
linked, and if any of those aren't upheld, then we don't reuse the candidate.
This should ensure the shared roots are identical to private roots.

Note that in practice this diff won't be enough to actually get much at all
sharing since there are a number of exclusions that need to be added to
the Omnibus environment to actually make this work.

#forcetdhashing

Reviewed By: ndmitchell

Differential Revision: D39099488

fbshipit-source-id: 223679475f611e0cea1ff044e74e6d0bc485a649