buck2: omnibus: annotate roots so we can tell why they are roots
Summary:
No functional change here, but this makes it easier to actually introspect
what Omnibus is doing and understand why a given node is being chosen as
a root.

Note that in an ideal world we have no implicit roots (which are roots as a
result of being depended on), and instead we only have explicit roots (which
aren't).

#forcetdhashing

Reviewed By: ndmitchell

Differential Revision: D39099489

fbshipit-source-id: 78faf00601475ec6a5423c8d414bd05aff80f9b4