buck2: add a supports_python_dlopen attribute to annotate Omnibus roots
Summary:
NOTE: This is all feature flagged out until D39168948

Like it says in the title. This will be a little more convenient than a
label in some cases.

This is a new attribute we're adding to allow C++ libraries to explicitly indicate that they'll be loaded by Python and consequently
need to be exposed as shared libraries.

This behavior is otherwise inferred at runtime (any C++ library that is depended on by Python is assumed to be dynamically
loaded). By making this explicit, we can make builds more efficient since we don't need to collect the full graph of Python
dependencies for a Python target in order to identify how specific C++ libraries will end up linked.

The goal long-term is to require this being explicit, but for now we're simply using this as a hint. If not set, the attribute may
still be inferred through Python -> C++ direct dependencies.

Note that it's always safe to set this to True even if it doesn't need to be (this is what we do in mode/dev). However, doing
so will make binaries slower to start if done too aggressively (e.g. setting it on _all_ libraries).

#forcetdhashing

Reviewed By: ndmitchell

Differential Revision: D39104635

fbshipit-source-id: 0a4813fb74034185e8b29cbb1555b2d72a9f2567