Check that declared external symlinks have existing remaining path
Summary:
External symlinks are a hack for us to figure out (later) what the external path the symlink is going to resolve to without actually reading through the symlink. They shouldn't be declared on materializer in the first place with a remaining path because those cannot be materialized (without first reading through the symlink). Add a check in all materializers that external symlinks with existing remaining paths being declared causes an error.

As a part of this change, now all materializers implement `declare_cas_many_impl` and `declare_copy_impl`, which gets called by `declare_cas_many` and `declare_copy` on `Materializer` trait.

Reviewed By: krallin

Differential Revision: D39142918

fbshipit-source-id: f7715d342737a2a69ddf8d543cb4b66d1a48f653