Setup protoc path for Cargo compilation in build.rs
Summary:
Set up the correct path to `protoc` in the `build.rs` script of the
proto crates. If `$PROTOC` is set nothing changes. If `$PROTOC` is not set,
check if the DotSlash path to `protoc` exists, if it does, use it for proto
compilation in tonic build.

Introduce a new `buck2_protoc_dev` crate that contains the helper and is used
as build dependency by the proto crates.

Reviewed By: krallin

Differential Revision: D39203860

fbshipit-source-id: d541b5066a0636e64542abaa7b4ec8956f586f52