ArtifactGroupValuesDyn
Summary:
This stack moves command executors to `buck2_execute`.

Executors depend on `ArtifactGroupValue`.

`ArtifactGroupValue` depends on `Artifact`. It probably shouldn't, but it does, and I have not found a good way to fix it.

`Artifact` depends on `Action`, so it can't be moved `buck2_execute`.

So `ArtifactGroupValuesDyn` trait provides access to subset of `ArtifactGroupValue` without depending on `Artifact` in the interface.

Reviewed By: krallin

Differential Revision: D39085028

fbshipit-source-id: 416144ea65b01a8dd24f29ae25821638a340aa3b