Modify get_or_declare_output to take in an optional set of associated artifacts
Summary:
Modify the `get_or_declare_output` signature to intake an optional IndexSet of ArtifactGroups. If it is set to Some of a value, that means the set should be added when creating a new StarlarkArtifact type. If there is already a StarlarkArtifact value, create a new one with its associate artifacts extended.

Next Up:
1. Migrate copy and symlink action to use this new parameter to pass on associated artifacts
2. Remove `get_bound_deprecated`

Reviewed By: bobyangyf

Differential Revision: D38501966

fbshipit-source-id: 6228977758e3ce8cc092007203d5733197b0a19c