Invalidate symlinks when changed
Summary:
Taken from: T128472966

When we read artifact metadata and traverse a symlink, we should track that we accessed this symlink so that if the symlink changes we invalidate reads that depended on it.

Here's we're retrieving the symlinks path (internal or external) and marking it as dependant so if the symlink changes we invalidate it and the files depending on it.

Reviewed By: krallin

Differential Revision: D38706127

fbshipit-source-id: 6e949c2fa927aef451370129f80e2f5435656c3e