Use json projections for javacd commands
Summary:
This changes the largest parts of the javacd json commands to be handled by tsets. This is a significant reduction to analysis time memory use (and should be similarly a reduction to execution time graph size).

source-only-abi actions need a filtered list of deps and that is a partly based on the target's own attrs and so still does a traversal and flattening. It's possible that we could propagate up basically a required-for-source-only-abi tset and merge that with the target's source_only_abi_deps and get the same behavior, but I'm not sure (and that should be a much smaller graph than the normal compilation graph).

Reviewed By: IanChilds

Differential Revision: D38953065

fbshipit-source-id: df150653736e2437e238638f3427d4909abafd83