buck2: replace `with_flushing(|client| client.stuff())` with `with_flushing().stuff()`
Summary:
Like it says in the title. The end goal here is to eliminate the closure, which
currently cannot capture anything from its context.

As of this diff, the logic that was in `with_flushing` is on the macros on
the `FlushingBuckdClient`. In an ideal world we'd have async destructors and
we would be able to actually do this in a `Drop` but right now we don't so
this is what it is.

Reviewed By: stepancheg

Differential Revision: D39206267

fbshipit-source-id: 531c71a1a236df389af995e48173ffaf034f3f13