buck2: buck2_client: implement Stdin shared abstraction
Summary:
Like it says in the title. We currently have something for Stdin in the LSP but I need this in more places,
so what I'd like to do is have a single Stdin we use everywhere, which means I need it to be shuffling bytes
and not decoded messages, which is what this does!

Reviewed By: stepancheg

Differential Revision: D39139164

fbshipit-source-id: 01ca901f284c35bf0dd5c5af722c47d956294d4a