buck2: use a macro for streaming methods
Summary:
I want to get rid of with_flushing because it prevents us from using anything
with a lifetime that isn't `'a` in that function, which is super constraining
and annoying.

So, I'm gonna put this logic inside the methods themselves, which will also
require less ceremony to call them.

To do this, I'm going to declare them using a macro, which is what this diff
does for streaming methods.

Reviewed By: stepancheg

Differential Revision: D39206261

fbshipit-source-id: d3a81059268fa93bb81f5cc5ad436e82c7064cbe