buck2: buck2_client: use new Stdin abstraction in LSP
Summary:
Like it says in the title, make this operate on a stream of Bytes instead of having direct access to stdin.

Prior to this diff, the `deserializer` could do blocking reads, now it cant, which is why we need a Decoder.

There's also a little annoying aspect in that Tonic wants a `'static` input stream, so we provide that. We could have
also allowed just `take`-ing the stdin, which would have worked, but I went a little step forward so that we could
imagine potentially releasing stdin and using it for something else after the LSP ops are done, even though we
probably will never ever do that.

Reviewed By: stepancheg

Differential Revision: D39139163

fbshipit-source-id: f8899df729a2d7dc934af22cbf56d7238b61a504