buck2: ignore WouldBlock in ConsoleInteractionStream
Summary:
This can happen if we are running `buck2 build | foo` and `foo` puts our stdin in
non-blocking mode (note: that wouldn't be `foo`'s stdin, but foo might be `nvim`
and might try to get access to the controlling tty).

For now, don't attempt to handle this properly (ideally we should _probably_ do
a `select` then attempt to read when we can, at which point we would likely get
SIGTTIN), but that's a lot of work for fairly little payoff, so let's just
ignore this when it happens and fallback to non-interactive.

Reviewed By: ndmitchell

Differential Revision: D39308401

fbshipit-source-id: 99f590d9756ff41157cf93ef56d3fa6b35e888f7