Remove table_builder::Push
Summary:
Previously to add a value to a "table", code needed to be written line this:

```
table.push().span(...)
```

`Push` could contain padding.

Now it is explicit:

```
table.rows.push(Row::span(...))
```

State is bad, functional is good.

This is not strictly necessary for the following diffs, but it helps. See the following diff D39287249.

Reviewed By: bobyangyf

Differential Revision: D39286461

fbshipit-source-id: 66800e634562f9ab4023fc8ef011911bff9c39e8