buck2: find_exclusions: omnibus_tools: ignore nodes that are shared_only but are also roots
Summary:
In Omnibus linking, if a node has shared_only = True, what we do is:

- If it's a root, nothing.
- If it's not a root, we exclude it (and transitively exclude all its
  dependents).

Update the find_exclusions script to do this. In theory, this shouldn't
really matter (it should be fine for us to choose to exclude a given library),
but in practice some libs are highly implementation-dependent (see D39200420,
D39215714) and excluding them causes runtime errors.

So, to maximize consistency with what Omnibus would *normally* do, have the
find_exclusions script ignore things that are annotated as Omnibus roots.

Note that of course, we're already being consistent assuming the
`supports_python_dlopen` annotations are consistent.

I think that long-term we should probably have the default behavior for
shared_only libs be to treat them as a root (thus leaving their dependents
alone), but that's potentially a big behavior change.

Reviewed By: ndmitchell

Differential Revision: D39256388

fbshipit-source-id: 3644c8542e8664b37304c7cc434550861a631a86