buck2: omnibus: give shared roots their own ldflags
Summary:
NOTE: This only takes effect when roots are shard, which is feature-flagged behind D39168948.

Like it says in the title. We can't really match the ldflags from the binaries
we receive, as there is too much variability (and most flags are relevant to
binaries not libs), but at the same time we do need them, in particular for
stripping and to keep GPU sections.

For stripping, I just get the default from Python. For GPU sections, we always
retain them, which is the right thing to do when we don't know whether or not
the lib will need them (note: the flags will still get applied to the libomnibus.so
as-is, which is what might actually need them in case it's too large to link
otherwise).

Note: this does mean that we don't honor strip_libpar = False on Python
libraries after this. I'll try to see if we can bring this back. That is
arguably better than not building them at all...

#forcetdhashing

Reviewed By: ndmitchell

Differential Revision: D39308069

fbshipit-source-id: d076e5d94726c237e073b494587024db75838562