Extract Xcode target creation
Summary:
# Overall Stack Context

We want to make it very easy to add a new Apple toolchains - for each new Xcode version, Pika version etc.

See D39275797 for final state for adding a new toolchain.

# Diff Context

The Xcode toolchain target logic was inlined in multiple places, so let's eliminate the duplication. This allows us to eliminate all the separate functions per SDK in a follow-up diff (functions removed in D39273957).

Reviewed By: rmaz

Differential Revision: D39273330

fbshipit-source-id: fa418622adfa21f6bdf0de2ed6f7917db524ac98