Extract generic toolchain arch selector
Summary:
# Overall Stack Context

We want to make it very easy to add a new Apple toolchains - for each new Xcode version, Pika version etc.

See D39275797 for final state for adding a new toolchain.

# Diff Context

We want to re-use constraint to architecture map for all toolchains - currently this logic is duplicated inline. This allows us to eliminate the Xcode-specific functions and unify behaviour (D39273957).

Reviewed By: rmaz

Differential Revision: D39273744

fbshipit-source-id: f08b155677051e04c75df8c68fbf3736e7dddb44