Define Apple toolchains in a generic way
Summary:
# Overall Stack Context

We want to make it very easy to add a new Apple toolchains - for each new Xcode version, Pika version etc.

See D39275797 for final state for adding a new toolchain.

# Diff Context

The logic for each toolchain was spread across multiple functions which made it harder to add. In this diff, we move all toolchain definitions in single data structure.

Reviewed By: rmaz

Differential Revision: D39274683

fbshipit-source-id: aa8813e24a9d0e72c643d22102096fed2157993a