Remove default arch from being passed
Summary:
# Overall Stack Context

We want to make it very easy to add a new Apple toolchains - for each new Xcode version, Pika version etc.

See D39275797 for final state for adding a new toolchain.

# Diff Context

The default architecture should not be retrieved twice and instead used in a single place. In the first step of this transformation, we remove the explicit passing of the architecture. Later on, it gets unified as we re-use a single arch select function (D39273957)

https://www.internalfb.com/code/fbsource/[d1ee2ee3d0c9dd120b31d881b0cd13880844de07]/fbcode/buck2/platform/toolchain/apple_toolchain_setup.bzl?lines=251-252

Reviewed By: rmaz

Differential Revision: D39273470

fbshipit-source-id: 2c4d55677e5884d63291b4a99dfae7a1d016c2e0