Bump max input file size to be sent to RE
Summary:
With RE now supporting large actions, let's raise the limit of this value gradually.
Based on a few shadow jobs there shouldn't be too much traffic, but we can get more T10/T6F machines if necessary.
Eventually, we should raise the limit for actions on RE to 500GB input files (the maximum value we can materialise on a T10).

Reviewed By: krallin

Differential Revision: D39345469

fbshipit-source-id: c0a5471d7ca65cdff7a53ba0bf2470a9df6cd419