Propagate link info for prebuilt_cxx_library deps
Summary:
For prebuilt libraries, sometimes deps are specified which include dependencies that must be linked along with the library itself.  For example, with LLVM:

https://www.internalfb.com/code/fbsource/[0283d94b8c38]/arvr/third-party/toolchains/platform010/build/llvm-fb/BUCK?lines=745-757

Without the `:LLVMSupport` dep, link errors are generated for missing symbols that `LLVMOption` requires.

To fix this, and match the buck1 behavior, propagate the link info from non-exported deps.

Reviewed By: christycylee

Differential Revision: D39264553

fbshipit-source-id: 234f03fa550524909709e1d61183f4c7e3044c62