Expose the info plist as a subtarget and it's path in XcodeData
Summary:
To run device builds via Xcode, Xcode needs to know the bundle identifier of the app. Typically, we make this customizable via variable substitution in plist generation, so to get the finalized bundle identifier, we need the generated plist.

To achieve this, we do two simple things:
1. expose the path of the generated plist in xcode data
2. add a subtarget to apple_bundle rules to materialize the plist file

This will allow us to read this file and extract any necessary information.

Reviewed By: milend

Differential Revision: D39222239

fbshipit-source-id: 8a5e929fd46f4ba284e4715ffe7bec7e503ba730