Add conversion of ArtifactMetadata to sqlite entry
Summary: Add an `ArtifactMetadataSqliteEntry` that represents the sqlite entry for the `ArtifactMetadata`. All of its fields implement rusqlite's FromSql trait so they can be inserted/queried directly with rusqlite. Add conversion methods between `ArtifactMetadata` and `ArtifactMetadataSqliteEntry`. The conversion from `ArtifactMetadataSqliteEntry` to `ArtifactMetadata` can throw errors, but these are all internal errors that we shouldn't be hitting in the first place (unless there's a sqlite schema change).  Add unit tests to verify that conversions shouldn't be hitting this.

Reviewed By: stepancheg

Differential Revision: D39229811

fbshipit-source-id: 79f6dd7f1942a0aad8cc381039b4c655ebce2687