Add a generic key-value sqlite table struct
Summary:
This diff starts adding the abstractions we need for storing materializer state in sqlite db by adding a struct for generic key-value sqlite table.
Keys and values are strings, and key is set as primary key.
This struct can be used later as table for storing version info or metadata.
Create table, insertions, and read all are all implemented and tested.
Having a struct for this is useful so that we can easily test functionality in isolation

Reviewed By: krallin

Differential Revision: D38775543

fbshipit-source-id: d33e5106fd04e40354fc52f5aefab53573b0800b