Concurrency handler to fully manage dice and updates
Summary:
More correct handling of concurrent commands so that it manages the updates to dice.

This way, we also ensure that when a command actually enters critical section, the DICE is initialized to the current state per disk instead of at the time the command is started.

This also ensures dice computes version in order.

This does introduce more churn where commands could be setting dice back and forth between two states, but concurrent commands should be minimal, and dice should be able to resurrect nodes when appropriate.

Reviewed By: krallin

Differential Revision: D39229060

fbshipit-source-id: 7a36ff9015f28941aa5b5fbdfca3c17a3b01e81c