buck2: create_unhashed_outputs: do cheap checks before expensive ones
Summary:
Its not very clear to me why we need this, but it looks like checking the count
of outputs on an artifact can take a while in aggregate. I guess this makes
sense since we can have large artifact groups as other-outputs?

Anyway, it's cheaper to check whether the artifact is indeed a default output
(which we care about) before attempting to iterate than it is to do the
reverse, so we might as well.

Reviewed By: nshipilov

Differential Revision: D39425072

fbshipit-source-id: 1897f16d60ea80583a227617d423823d0fe8af1d