buck2: do an atomic rename when writing `--out` to a file
Summary:
This is very much a papercut that I only care about, but I often do this:

```
 buck2 build fbcode//mode/opt fbcode//buck2/cli:buck2 --out ~/buck2
```

That doesn't work if `buck2` is currently running in a daemon somewhere because
then you can't open it for writing (as it's ETXTBUSY), so you can't open it for
writing. You can however do operations that replace the file entirely, however,
like moving over it.

So, this diff fixes this. Note that I only use this mechanism as error-recovery
since it wouldn't necessarily work if you tried to e.g. write to `/dev/stdout`.
I also added tests for that behavior that I didn't break.

Reviewed By: stepancheg

Differential Revision: D39385069

fbshipit-source-id: 82a08a4679666af9a8805271d490ef0bc53a43f3