Add a test for the oncall function
Summary: Checks it works, and then appears in the appropriate output. I could have created an isolated test, but we should oncall our TARGETS anyway, so this seemed easier. I put both the cquery and uquery test in the same file, because to split them would make a lot less sense.

Reviewed By: stepancheg

Differential Revision: D39325289

fbshipit-source-id: 6a9e5926e1bbed619bb66005113889c3835197e5