buck2: add a reduction for external_debug_info presence
Summary:
D39241134 (https://github.com/facebookincubator/buck2/commit/e87c37237c1bce0ebe68d1776d3778261bddf29b) adds a lot of hidden outputs that are all effectively empty, but
require traversing to find that out. For example, this is the shape of hidden
outputs for a Python binary that this produces:

```
[[], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], [], []]
```

Notionally, this is empty, but you need to walk the whole thing to figure this
out.

In builds we end up deduplicating materializations and we have concurrency, but
output symlinks don't do that, so this created a bunch of slowness in that
step:

- https://fb.workplace.com/groups/buck2users/permalink/3234838533439172/
- https://fb.workplace.com/groups/buck2users/permalink/3232690260320666/
- https://fb.workplace.com/groups/buck2users/permalink/3232782826978076/

I've already submitted a workaround in D39425072 (https://github.com/facebookincubator/buck2/commit/d677e41253b73a31aafa1255a532c38992482efd), but it seems like a
reasonable thing to do to also update the code here to avoid doing all this
extra work to begin with.

An alternative here would be for this info to live in its own tset rather than
in the LinkInfo tset, which would allow for nodes with no value.

Differential Revision: D39463080

fbshipit-source-id: 43e7d10d1acb513f5d2e2dd78a9e614d791215ef