Add sqlite db reading and writing
Summary:
Hook up `MaterializerStateSqliteDb` to `DeferredMaterializer`.

Now `DeferredMaterializer` accepts 2 more parameters: one for the sqlite db and one for the loaded materializer state. It initializes entries in the `ArtifactTree` based on the loaded materializer state. On declare, invalidation, and materialization failures, we delete paths from sqlite db. On materialization success, we insert a path into the sqlite db.

Reviewed By: krallin

Differential Revision: D39372432

fbshipit-source-id: 905ac9685dfb66e7fdea9e36c9329876fab28a85