Allow null values in key-value sqlite table
Summary: Certain keys in our metadata sqlite table may have null values because they are not set. An example is the buckconfig sqlite version if there is no such buckconfig. In these cases, we'd like to differentiate between empty string and unset, so allow our key-value sqlite tables to have null values (but not null keys)

Reviewed By: krallin

Differential Revision: D39414130

fbshipit-source-id: 48f61ec2ab048e8db58e7f7d58ae481173b0c097