Add QueryValueSet
Summary:
We probably want quite a few functions that operate only on set-like things, e.g. filter, so define a query argument that has that restriction.

This type is isomorphic to the type QueryEvaluationValue, but you can reasonably imagine that one way we extent QueryEvaluationValue, so I thought it best to keep them separate.

Reviewed By: bobyangyf

Differential Revision: D39521725

fbshipit-source-id: dc33125d1d487ab3200105ecbddc4f2617dcd8e5